+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 * Modify AEM Author instance's https listen port from 5433 to 5432
 * Enable AutoScalingGroups metrics collection #56
 * Move aem_password_reset_version, aem_orchestrator_version, oak_run_version stack facts to stack-provisioner-hieradata config #11
+* Modify cloud-init's extra local.yaml config to be appended to base local.yaml
 
 ### 2.0.0
 * Add Stack Provisioner custom hiera configuration support